test: Injector Test 작성
