+++ README.md
@@ -8,7 +8,7 @@
 다음 문제점을 해결한다.
 - [x] 의존성 주입이 필요한 필드와 그렇지 않은 필드를 구분할 수 없다.
   - [x] Annotation을 붙여서 필요한 요소에만 의존성을 주입한다.
-  - [ ] 내가 만든 의존성 라이브러리가 제대로 작동하는지 테스트 코드를 작성한다.
+  - [x] 내가 만든 의존성 라이브러리가 제대로 작동하는지 테스트 코드를 작성한다.
 ### Recursive DI
 - [x] CartRepository가 다음과 같이 DAO 객체를 참조하도록 변경한다.
 CartProductEntity에는 createdAt 프로퍼티가 있어서 언제 장바구니에 상품이 담겼는지를 알 수 있다.
